refactor(arrow): export the ColReader wrapper for ArrowColReader
This was previously an internal package but it would be useful in the
other projects during the migration.
